pyproject.toml
[project]
name = "multiaiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "jinja2>=3.1.6",
    "jupyter>=1.0.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.5",
    "langsmith>=0.4.14",
    "lxml>=4.9.0",
    "markdown>=3.8.2",
    "matplotlib>=3.7.0",
    "notebook>=7.0.0",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "pillow>=11.3.0",
    "plotly>=5.17.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.1.1",
    "pytz>=2023.3",
    "reportlab>=4.4.3",
    "requests>=2.31.0",
    "scipy>=1.11.0",
    "seaborn>=0.12.0",
    "selenium>=4.15.0",
    "streamlit>=1.48.1",
    "tavily-python>=0.3.0",
]
